initial the dialog component
