+++ src/client/src/component/Cart/Cart.styles.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 export const Wrapper = styled.aside`
-  font-family: Arial, Helvetica, sans-serif;
-  width: 500px;
-  padding: 20px;
+	font-family: Arial, Helvetica, sans-serif;
+	width: 500px;
+	padding: 20px;
 `;
