+++ src/client/src/component/CheeseCard/component/DetailDialog.styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+
+export const Wrapper = styled.div`
+
+`;
