+++ src/client/src/component/CheeseCard/component/DetailDialog.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const DetailDialog: React.FC = ()=> {
+	return(
+		<div>
+			
+		</div>
+	)
+}
+
+export default DetailDialog;
